% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ebm.R
\name{fit_ebm}
\alias{fit_ebm}
\title{Find maximum likelihood estimate}
\usage{
fit_ebm(data, p_vec, clinical_info, nstart, initial_iter)
}
\arguments{
\item{data}{A data frame.}

\item{p_vec}{Unique row identifier.}

\item{nstart}{Number of initialization points}

\item{initial_iter}{Number of iterations}

\item{info}{A data frame}
}
\value{
A list
}
\description{
Find maximum likelihood estimate
}
