% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_STATIS.R
\name{fit_STATIS}
\alias{fit_STATIS}
\title{Fit TB-STATIS}
\usage{
fit_STATIS(data, p_vec, clinical_info, nstart, initial_iter)
}
\arguments{
\item{data}{A data frame of observed data.}

\item{p_vec}{A vector of clinical measure accuracies.}

\item{clinical_info}{A data frame of clinical measures and states.}

\item{nstart}{Number of sequence initializations for maximum likelihood search.}

\item{initial_iter}{Number of iterations for maximum likelihood search.}
}
\value{
A list of results.
}
\description{
Estimate maximum likelihood sequence for TB-STATIS. User must input the observed data,
set of clinical states and measures to include in the model, and a vector of probabilities indicating the accuracy
for each clinical measure.'fit_STATIS' returns the maximum likelihood sequence, and all log-likelihoods calculated during
estimation.
}
