% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_group}
\alias{get_group}
\title{Propose simultaneous clinical states}
\usage{
get_group(info = data.frame(), all_seqs)
}
\arguments{
\item{info}{A data frame.}

\item{all_seqs}{Data frame of all possible disease sequences.}
}
\value{
A data frame.
}
\description{
Generate a simulataneous disease sequence given a sequence of clinical states.
}
\keyword{internal}
