% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_likelihood.R
\name{get_likelihood}
\alias{get_likelihood}
\title{Calculate log-likelihood}
\usage{
get_likelihood(data = data.frame(), S = data.frame(), p_vec = vector())
}
\arguments{
\item{data}{A data frame.}

\item{S}{Data frame}

\item{p_vec}{A vector}
}
\value{
A list.
}
\description{
Calculate log-likelihood
}
