% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_seq}
\alias{get_seq}
\title{Sequence initialization}
\usage{
get_seq(info = data.frame())
}
\arguments{
\item{info}{A data frame containing the clinical measures and corresponding clinical
states to generate a sequence from.}
}
\value{
A data frame.
}
\description{
Generate initial sequence as start point for maximum likelihood search.
}
\keyword{internal}
