% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stage.R
\name{get_stage}
\alias{get_stage}
\title{Get disease stage}
\usage{
get_stage(data = data.frame(), S = data.frame(), p = vector())
}
\arguments{
\item{data}{A data frame.}

\item{S}{Data frame}

\item{p}{A vector}
}
\value{
A data frame.
}
\description{
Get disease stage
}
