% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_class}
\alias{plot_class}
\title{Disease class distribution}
\usage{
plot_class(pred_class = vector())
}
\arguments{
\item{pred_class}{A vector of predicted disease classes.}
}
\value{
A ggplot object.
}
\description{
Generate a bar plot of the disease class distribution using the vector containing the predicted
disease class for each individual after fitting TB-STATIS.
}
