% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_events}
\alias{plot_events}
\title{Plot clinical events}
\usage{
plot_events(data = data.frame(), id.var = character(), event.vars = vector())
}
\arguments{
\item{data}{A data frame.}

\item{id.var}{Unique row identifier.}

\item{event.vars}{Variable names for clinical events.}
}
\value{
A ggplot object.
}
\description{
Plot clinical events
}
