% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_likes}
\alias{plot_likes}
\title{Plot likelihood ascent}
\usage{
plot_likes(likes = data.frame())
}
\arguments{
\item{likes}{A data frame.}
}
\value{
A ggplot object.
}
\description{
Plot likelihood ascent
}
