% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{possible_seqs}
\alias{possible_seqs}
\title{All possible sequences}
\usage{
possible_seqs(N = numeric())
}
\arguments{
\item{N}{Number of clinical states.}
}
\value{
A data frame.
}
\description{
Generate data frame of all possible simultaneous sequences for N clinical states.
}
\keyword{internal}
