% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_class.R
\name{pred_class}
\alias{pred_class}
\title{Get disease class}
\usage{
pred_class(data = data.frame(), S = data.frame(), p = vector())
}
\arguments{
\item{data}{A data frame.}

\item{S}{Data frame}

\item{p}{A vector}
}
\value{
A data frame.
}
\description{
Get disease class
}
