% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvd.R
\name{pvd}
\alias{pvd}
\title{Plot positional variance diagram}
\usage{
pvd(boot_seqs = data.frame(), ml = data.frame())
}
\arguments{
\item{boot_seqs}{A vector.}

\item{ml}{description}
}
\value{
A ggplot object.
}
\description{
Plot positional variance diagram
}
