% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TBData}
\alias{TBData}
\title{TBData}
\format{
A data frame with 250 rows and 8 columns:
\describe{
  \item{Index}{unique row identifier}
  \item{class}{disease class}
  \item{X1}{Clinical measure 1, state 1}
  \item{X2}{Clinical measure 1, state 2}
  \item{X3}{Clinical measure 2, state 1}
  \item{X4}{Clinical measure 2, state 2}
  \item{X5}{Clinical measure 2, state 3}
  \item{X6}{Clinical measure 3, state 1}
  \item{C1}{Covariate, categorical}
  \item{C2}{Covariate, continuous}
}
}
\usage{
TBData
}
\description{
A simulated data set to fit TB-STATIS. The data contains 3 clinical measures with 2,3,and 1 events respectively for a total of 6 disease events.
}
\keyword{datasets}
