% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TBebmData}
\alias{TBebmData}
\title{TBebmData}
\format{
A data frame with 250 rows and 8 columns:
\describe{
  \item{Index}{unique row identifier}
  \item{stage}{true disease stage}
  \item{V1}{Clinical measure 1, level 1}
  \item{V2}{Clinical measure 1, level 2}
  \item{V3}{Clinical measure 2, level 1}
  \item{V4}{Clinical measure 2, level 2}
  \item{V5}{Clinical measure 2, level 3}
  \item{V6}{Clinical measure 3, level 1}
}
}
\usage{
TBebmData
}
\description{
A simulated data set to fit the EBM. The data contains 3 clinical measures with 2,3,and 1 events respectively for a total of 6 disease events.
}
\keyword{datasets}
