% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_seq.R
\name{bootstrap_seq}
\alias{bootstrap_seq}
\title{Find maximum likelihood estimate}
\usage{
bootstrap_seq(z, data, clinical_info, p_vec, nstart, initial_iter)
}
\arguments{
\item{z}{number of bootstrap samples}

\item{data}{A data frame.}

\item{p_vec}{Unique row identifier.}

\item{nstart}{Number of initialization points}

\item{initial_iter}{Number of iterations}

\item{info}{A data frame}
}
\value{
A list
}
\description{
Find maximum likelihood estimate
}
