% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_STATIS.R
\name{fit_STATIS}
\alias{fit_STATIS}
\title{Find maximum likelihood estimate}
\usage{
fit_STATIS(data, p_vec, clinical_info, nstart, initial_iter)
}
\arguments{
\item{data}{A data frame.}

\item{p_vec}{Unique row identifier.}

\item{clinical_info}{A data frame}

\item{nstart}{Number of initialization points}

\item{initial_iter}{Number of iterations}
}
\value{
A list
}
\description{
Find maximum likelihood estimate
}
