% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_group}
\alias{get_group}
\title{Propose simultaneous events}
\usage{
get_group(info = data.frame(), all_seqs = data.frame())
}
\arguments{
\item{info}{A data frame.}

\item{all_seqs}{List of all possible sequences.}
}
\value{
A ggplot object.
}
\description{
Propose simultaneous events
}
\keyword{internal}
