% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_seq}
\alias{get_seq}
\title{Propose sequence as start point for maximum likelihood estimation.}
\usage{
get_seq(info = data.frame())
}
\arguments{
\item{info}{A data frame.}
}
\value{
A data frame.
}
\description{
Propose sequence as start point for maximum likelihood estimation.
}
\keyword{internal}
