% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_states}
\alias{plot_states}
\title{Plot clinical states}
\usage{
plot_states(data = data.frame(), id.var = character(), state.vars = vector())
}
\arguments{
\item{data}{A data frame.}

\item{id.var}{Unique row identifier.}

\item{state.vars}{Variable names for clinical states.}
}
\value{
A ggplot object.
}
\description{
Plot clinical states
}
