% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{possible_seqs}
\alias{possible_seqs}
\title{Generate data frame of all possible simultaneous sequences.}
\usage{
possible_seqs(N = numeric())
}
\arguments{
\item{N}{Number of events.}
}
\value{
A data frame.
}
\description{
Generate data frame of all possible simultaneous sequences.
}
\keyword{internal}
