% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvd.R
\name{pvd}
\alias{pvd}
\title{Positional variance diagram}
\usage{
pvd(boot_seqs = data.frame(), ml = data.frame())
}
\arguments{
\item{boot_seqs}{A data frame of estimated sequences obtained from bootstrap resampling and
estimating TB-STATIS on each resample.}

\item{ml}{Maximum likelihood sequence from fitting TB-STATIS to original data.}
}
\value{
A ggplot object.
}
\description{
Plot maximum likelihood sequence with sequences fit using bootstrap resampling to visualize
uncertainty about maximum likelihood sequence.
}
